Fixes (1)
Fixed sign position on the antechamber
Fixed first teleport text
Fixed antechamber destination position
Tweaked nacelle inside destination rotation
Now the player rotation is set on teleportation

+++ Assets/Scripts/Object/Info/InfoController.cs
@@ -37,7 +37,7 @@ namespace WindTurbineVR.Object
 
         [SerializeField] protected GameObject prefabUI;
 
-        [SerializeField] protected int size = 1;
+        [SerializeField] protected float size = 1;
 
         [SerializeField] private GameObject uiInstance;
